+++ src/data/index.ts
@@ -6,6 +6,7 @@ import { tabatas } from './tabatas';
 import { workouts } from './workouts';
 import { beeps } from './beeps';
 import { pomodoros } from './pomodoros';
+import { meditations } from './meditations';
 
 export const timerPages: ProgrammaticPage[] = [
   ...countdowns,
@@ -14,6 +15,7 @@ export const timerPages: ProgrammaticPage[] = [
   ...workouts,
   ...beeps,
   ...pomodoros,
+  ...meditations,
 ];
 
 export const pageByPath = (path: string): ProgrammaticPage | undefined =>